loader: GH1962; Fix for assembler when compiling 32-bit
try_compile was always failing for 32 bit builds. The cause was the
linker was being passed the same flags as the assebler but the linker
requires a different flags for 32 bit.

This fix has custom CMake modules that are now located in the
loader/cmake folder, as well as some changes to loader/CMakeList.txt.