layers: Add warning for LOAD_OP_LOAD+LAYOUT_UNDEFINED
While not technically an error (VK_IMAGE_LAYOUT_UNDEFINED is always
allowed as an initial layout), this combination is almost never what
you actually want.  If nothing else, it's a potential performance
problem because VK_ATTACHMENT_LOAD_OP_LOAD is liable to be more
expensive than VK_ATTACHMENT_LOAD_OP_DONT_CARE.  Give developers a
helpful warning in this case.

This would have caught an actual bug in some of the Sascha Willems
demos as fixed in this PR:

https://github.com/SaschaWillems/Vulkan/pull/400

Change-Id: I73c64ae60a65cfb200db00707d650ec8da7ba130