layers:Add standard validation helper header
This header captures the standard validation layer stack in a C array
and a C++ std::vector.
Instructions for its use are included in the header comment.
It's intended to simplify enabling the standard validation layer stack
and unifying how that stack is enabled regardless of which loader an
app is using.